requirements.txt
# requirements.txt
deepl==1.16.1
python-dotenv==1.0.0

# For development:
pytest==7.4.3
black==23.11.0
flake8==6.1.0
mypy==1.7.0
bandit==1.7.5